intelligent_document_ingestion/file-processing-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
boto3
pydantic
pypdf2
pytesseract
Pillow
python-docx
pandas
